[EDITED/NavigationLink] Now is more reusable. Moved to the components folder
